% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simField.R
\name{simField}
\alias{simField}
\title{Simulate A continous 2D spatial field evolving over time}
\usage{
simField(M, rangeE = 0.8, sigmaE = 0.04, rhoAR = 0.9,
  sigmaRW = 0.001, ageB = arm::logit((7:1/1000) + 0.00025),
  deltaUrban = -0.1, offset = 1, max.edge = c(0.15, 0.3))
}
\arguments{
\item{M}{Number of years}

\item{rangeE}{spatial range}

\item{sigmaE}{spatial variance}

\item{rhoAR}{Yearly autoregression}

\item{sigmaRW}{Variation of independent temporal component}

\item{ageB}{age specfifc effects}

\item{deltaUrban}{urban effect}

\item{offset}{mesh creation parameter}

\item{max.edge}{mesh creation parameter}
}
\value{
field object, contains multiple items. Spatial points data frame with 
raster values for transformed linear combination, and other values. A mesh 
that was used to create the latent field. The latent field
itself. A bounding shape where all observations take place. A projection 
matrix from the mesh to the entire field of interest. The spde for the matern
approximation.
}
\description{
Calculates a continous underlying spatial field for the
Dominican Republic over time and across ages for probability of mortality
for one of 7 age groups. The final observed field is a linear additive model
as described in Wakefield 2017.
}
\examples{
\dontrun{
require(tidyverse)

fieldDR <- simField(4, rangeE = 1.3)
 
fieldDR$fqzDF \%>\%
    ggplot(aes(x=long, y=lat, fill=fqz)) +
    geom_raster() +
    facet_wrap(~year) +
    coord_equal() +
    theme_void() +
    scale_fill_distiller(palette = "Spectral")
}

}
